P05/automation/requirements.txt
pysnmp==4.4.12
PyYAML==6.0.1
